Returned the Map from fetchDetails function in AppwriteService
Used Map instead of Array, which was used previously.
The keys in the map will be the youtube video ids.
It makes it faster to retrive the data from a perticular id very fast.
Also, It helps to preserve the order, since the original ids can be mapped to its details using Map as data warehouse.

Also updated the testing snippet in App.jsx to use the Map and retain the order of recommendation