Created the Recommendation class in main dataManager script file
init function in Recommendation class fetches the recommendation collection from appwrite, flattens it and stores it in a private variable.
Created a getter method which returns the copy of private variable.

Also added and used the utility functions to shuffle an array.

Updated the App.jsx to test it out